+++ 2.2_discount_codes.sql
@@ -47,7 +47,7 @@ DECLARE
     v_discount_record RECORD;
     v_discount_amount INT := 0;
     
-    v_user_balance INT;
+    v_user_balance NUMERIC(12,1);
     v_inventory_ids UUID[];
     v_contents TEXT[];
     
@@ -146,11 +146,11 @@ BEGIN
     -- 1. Deduct Points (If total_price > 0)
     IF v_total_price > 0 THEN
         DECLARE
-            v_current_bonus INT;
-            v_current_paid INT;
-            v_deduct_bonus INT := 0;
-            v_deduct_paid INT := 0;
-            v_remaining_cost INT := v_total_price;
+            v_current_bonus NUMERIC(12,1);
+            v_current_paid NUMERIC(12,1);
+            v_deduct_bonus NUMERIC(12,1) := 0;
+            v_deduct_paid NUMERIC(12,1) := 0;
+            v_remaining_cost NUMERIC(12,1) := v_total_price;
         BEGIN
             SELECT bonus_balance, paid_balance INTO v_current_bonus, v_current_paid
             FROM points_balance WHERE user_id = p_user_id;
