+++ commercial_points_functions.sql
@@ -17,8 +17,8 @@ LANGUAGE plpgsql
 SECURITY DEFINER
 AS $$
 DECLARE
-    new_paid INT;
-    new_bonus INT;
+    new_paid NUMERIC(12,1);
+    new_bonus NUMERIC(12,1);
 BEGIN
     -- 1. Validation
     IF p_amount <= 0 THEN
@@ -58,10 +58,10 @@ SECURITY DEFINER
 AS $$
 DECLARE
     target_user_id UUID := auth.uid();
-    current_bonus INT;
-    current_paid INT;
-    deduct_from_bonus INT := 0;
-    deduct_from_paid INT := 0;
+    current_bonus NUMERIC(12,1);
+    current_paid NUMERIC(12,1);
+    deduct_from_bonus NUMERIC(12,1) := 0;
+    deduct_from_paid NUMERIC(12,1) := 0;
 BEGIN
     -- 1. Lock record for update
     SELECT bonus_balance, paid_balance INTO current_bonus, current_paid
