+++ supabase/fn_admin_deduct_points_site.sql
@@ -14,12 +14,12 @@ LANGUAGE plpgsql
 SECURITY DEFINER
 AS $$
 DECLARE
-    current_bonus INT;
-    current_paid INT;
-    deduct_from_bonus INT := 0;
-    deduct_from_paid INT := 0;
-    actual_deducted INT := 0;
-    existing_amount INT := 0;
+    current_bonus NUMERIC(12,1);
+    current_paid NUMERIC(12,1);
+    deduct_from_bonus NUMERIC(12,1) := 0;
+    deduct_from_paid NUMERIC(12,1) := 0;
+    actual_deducted NUMERIC(12,1) := 0;
+    existing_amount NUMERIC(12,1) := 0;
 BEGIN
     IF p_reference_id IS NOT NULL THEN
         SELECT ABS(amount) INTO existing_amount
